0011198: ITK code is inconsistent include usage
There is an inconsistent usage of #include <> vs #include ""

Aside from being inconsistent from a stylistic point of view, this
requires me to configure of project such that both user and system
include paths point to ITK. This slows down builds

This was discussed on the mailing list on 2010-06-22.

There are hundreds of files that need to change for this to be made
self-consistent.

Fixed looking for <itk.*> to change to "itk.*"
Fixed vxl, vnl, vcl include files to use ""
Change include for Utilities directory
Changed system level to use <>

Change-Id: I9fc90e4fad30de982299d8955fbc33193713a503